one more minor change to fix the bug -- BGG, 5/16/12

+++ src/rtcmix/minc_functions.cpp
@@ -512,7 +512,7 @@ double m_trand(float p[], int n_args, double pp[])
 	int trunc = m_irand(p, n_args, pp);
 
 	// this is to make the lower bound excluded for negative lower bound
-	if (n_args == 2 && p[0] != p[1])
+	if (n_args == 2 && (int)p[0] != (int)p[1])
 		while (trunc == (int)p[0])
 			trunc = m_irand(p, n_args, pp);
